fix: widen the three average holding periods to decimal
`HolderPerformance.AverageHoldingPeriod` and its Top10/Top20 siblings were
`int?` on the strength of a measurement that asked the wrong question.

Measured 2026-08-29 across 391 rows from five large filers (Berkshire 53,
BlackRock 71, Vanguard 104, State Street 53, FMR/Fidelity 110): all 1,173
values across the three fields arrive as bare JSON integers — none
fractional, none quoted, none carrying a decimal point. Verified two ways,
by parsed JSON type and by raw-text scan for a decimal point. `int?` is
correct for the data FMP serves today, and that is not what changed.

What the original measurement missed is that `int` rejects representations,
not just values. Measured against this library's own FmpJsonContext options,
`int?` throws on `20.0`, `20.5`, `"20.0"`, `"20.5"` and `1e2`, while
`decimal?` accepts all five; the context's AllowReadingFromString rescues
quoted *integers* only. FMP therefore does not have to change these numbers
to break the binding — only print them differently, which is a serializer
setting on their side rather than a data change.

The blast radius is the whole response, also measured: a three-row array
with one offending value binds 3 rows as `decimal?` and 0 as `int?`. The
rows that parsed are lost with the one that did not, so a single `20.0`
costs a caller all 53 of Berkshire's rows.

Widening costs a less tidy type on a value that is always whole. Not
widening costs the entire response. Only these three fields move: they are
the only means on the record. The genuine counts — portfolioSize,
securitiesAdded, securitiesRemoved, holdingPeriod, year, quarter — stay
`int?`, because there the type is right on its own terms.

The new test pins the wire form rather than the data, mirroring
An_unparseable_quoted_number_costs_one_field_not_the_row. Reintroducing
`int?` now fails at compile time, not at runtime.

CONTRIBUTING's typing convention is amended in the same commit, because as
written it is what produced this: "unless you have measured that it cannot
be fractional" is not the test that keeps a binding safe.

Done while the change is free: no tags, no releases, VersionPrefix 0.1.0,
unpublished. After publish this is a breaking change to public API.

+++ src/FmpDotNet/Models/HolderSummaries.cs
@@ -75,7 +75,7 @@ public sealed record HolderIndustryBreakdown
 ///
 /// <para><b>Three fields here are genuine counts and are <see cref="int"/>:</b>
 /// <see cref="PortfolioSize"/>, <see cref="SecuritiesAdded"/> and <see cref="SecuritiesRemoved"/>. The three
-/// average holding periods are means rather than counts but are <see cref="int"/> as well — see
+/// average holding periods are means rather than counts and are <see cref="decimal"/> — see
 /// <see cref="AverageHoldingPeriod"/> for the measurement behind that. Everything else is money or a percentage
 /// and is <see cref="decimal"/> — see <see cref="HolderAnalytics"/> for why.</para></summary>
 public sealed record HolderPerformance
@@ -117,20 +117,30 @@ public sealed record HolderPerformance
 
     /// <summary>The mean number of quarters the filer has held its positions.
     ///
-    /// <para><b>A mean, not a count — <see cref="int"/> anyway.</b> Measured 2026-08-29 across 391 rows from
-    /// five large filers (Berkshire Hathaway 53, BlackRock 71, Vanguard 104, State Street 53, FMR/Fidelity 110):
-    /// across this field and its two siblings below, 0 of 1,173 values were fractional. FMP rounds them. These
-    /// three are the only means on this record's <see cref="int"/> list, so if FMP ever stops rounding one, that
-    /// single value costs the caller the whole response.</para></summary>
-    [JsonPropertyName("averageHoldingPeriod")] public int? AverageHoldingPeriod { get; init; }
+    /// <para><b>A mean, not a count, and <see cref="decimal"/> even though every value measured was whole.</b>
+    /// Measured 2026-08-29 across 391 rows from five large filers (Berkshire Hathaway 53, BlackRock 71,
+    /// Vanguard 104, State Street 53, FMR/Fidelity 110): across this field and its two siblings below, all
+    /// 1,173 values arrived as bare JSON integers — none fractional, none quoted, none carrying a decimal
+    /// point. FMP rounds these, and writes them without a point.</para>
+    ///
+    /// <para><b>The type follows what being wrong would cost, not what was seen.</b> <see cref="int"/> rejects
+    /// <c>20.0</c> exactly as hard as <c>20.5</c>: measured 2026-08-29 against this library's own
+    /// <c>FmpJsonContext</c> options, all of <c>20.0</c>, <c>20.5</c>, <c>"20.0"</c>, <c>"20.5"</c> and
+    /// <c>1e2</c> throw, because the context's <c>NumberHandling = AllowReadingFromString</c> rescues quoted
+    /// <i>integers</i> only. So FMP would not have to change this number to break the binding — it would only
+    /// have to print it differently, which is a serializer setting on their side rather than a data change.
+    /// And the throw is not confined to the offending row: one bad value aborts the whole array, so a single
+    /// <c>20.0</c> costs the caller all 53 of Berkshire's rows including the 52 that parsed. Widening costs a
+    /// less tidy type on a value that is always whole; not widening costs the entire response.</para></summary>
+    [JsonPropertyName("averageHoldingPeriod")] public decimal? AverageHoldingPeriod { get; init; }
 
     /// <summary>The same, over the ten largest positions. See <see cref="AverageHoldingPeriod"/> for the
-    /// measurement behind the <see cref="int"/> typing.</summary>
-    [JsonPropertyName("averageHoldingPeriodTop10")] public int? AverageHoldingPeriodTop10 { get; init; }
+    /// measurement behind the <see cref="decimal"/> typing.</summary>
+    [JsonPropertyName("averageHoldingPeriodTop10")] public decimal? AverageHoldingPeriodTop10 { get; init; }
 
     /// <summary>The same, over the twenty largest. See <see cref="AverageHoldingPeriod"/> for the measurement
-    /// behind the <see cref="int"/> typing.</summary>
-    [JsonPropertyName("averageHoldingPeriodTop20")] public int? AverageHoldingPeriodTop20 { get; init; }
+    /// behind the <see cref="decimal"/> typing.</summary>
+    [JsonPropertyName("averageHoldingPeriodTop20")] public decimal? AverageHoldingPeriodTop20 { get; init; }
 
     /// <summary>Portfolio turnover for the quarter, as a fraction.</summary>
     [JsonPropertyName("turnover")] public decimal? Turnover { get; init; }
